Simplify tests with fixtures
